// src/components/proveedores/types.tsx
export interface ProveedorForm {
    codigoProveedor: string;
    nombreEmpresa: string;
    ruc: string;
    contacto: string;
    telefono: string;
    correo: string;
    direccion: string;
    ciudad: string;
    pais: string;
    tipoProveedor: string;
    terminosPago: string;
    observacion: string;
}